[enhancement] Changed information about committing changes and encountering issues in setup
Changed commits from "a different branch" to "your fork" as better represents GitHub and the process needed to be taken. Also tells them about Travis/coveralls and to be patient when waiting for a review.

When an error is encountered during installation, they should ask via communication channels, create an issue for it then are encouraged to try and fix it themselves.